files/requirements.txt
# все пакеты либо уже стоят в Colab, либо лёгкие
feedparser
requests
python-dotenv
pandas==2.2.2      # совпадает с Colab
matplotlib==3.8.4  # совместимо с pandas-2.2
nltk==3.9.1
tqdm
